+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\PC_User\.android\avd\Nexus_5X_API_28.avd" />
+            <value value="C:\Users\PC_User\.android\avd\Pixel_4_API_30.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-07-27T08:15:43.194016600Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-10-14T05:50:38.671899200Z" />
   </component>
 </project>
