+++ .idea/misc.xml
@@ -11,7 +11,8 @@
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/activity_main.xml" value="0.15" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/bottom_sheet.xml" value="0.33" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/bottom_sheet_contents.xml" value="0.22" />
-        <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/fragment_home.xml" value="0.1" />
+        <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/fragment_detail.xml" value="0.15" />
+        <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/fragment_home.xml" value="0.22" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/fragment_main.xml" value="0.1" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/fragment_search_bottom_sheet_dialog.xml" value="0.18" />
         <entry key="..\:/Users/PC_User/AndroidStudioProjects/Tanmen/app/src/main/res/layout/fragment_shuffle.xml" value="0.1" />
