+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.android.example.tanmen"
+        applicationId "jp.example.tanmen"
         minSdk 23
         targetSdk 31
         versionCode 1
