--- app/src/main/java/com/android/example/tanmen/Model/Shop.kt
@@ -1,10 +1,0 @@
-package com.android.example.tanmen.Model
-
-import java.io.Serializable
-
-data class Shop(
-    var image: String,
-    var name: String,
-    var address: String?,
-    var hours: String?
-): Serializable
