+++ app/src/main/java/jp/example/tanmen/Controller/Activity/MainActivity.kt
@@ -1,0 +1,96 @@
+package jp.example.tanmen.Controller.Activity
+
+import android.Manifest
+import android.content.Context
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.location.Location
+import android.location.LocationListener
+import android.location.LocationManager
+import android.os.Bundle
+import android.provider.Settings
+import android.util.Log
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.core.app.AppLaunchChecker
+import androidx.core.content.ContextCompat
+import jp.example.tanmen.API.ShopService
+import jp.example.tanmen.databinding.ActivityMainBinding
+
+class MainActivity : AppCompatActivity(), LocationListener {
+    private lateinit var binding: ActivityMainBinding
+    private var requestPermissionLauncher = registerForActivityResult(
+        ActivityResultContracts.RequestPermission()
+    ) { isGranted: Boolean ->
+        if (isGranted) {
+            locationStart()
+            Log.d("location", "許可されました。")
+        } else {
+            Log.d("location", "拒否されました。")
+        }
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        var boot = AppLaunchChecker.hasStartedFromLauncher(applicationContext)
+        if (!boot) {
+            AlertDialog.Builder(this)
+                .setMessage("右下の検索ボタンから検索してください。")
+                .setPositiveButton("Ok", null)
+                .show()
+
+            AppLaunchChecker.onActivityCreate(this)
+        }
+
+        if (ContextCompat.checkSelfPermission(this,
+                Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+            requestPermissionLauncher.launch(Manifest.permission.ACCESS_FINE_LOCATION)
+        } else {
+            locationStart()
+        }
+    }
+
+    private fun locationStart() {
+        Log.d("locationStart()", "locationStart()")
+
+        val locationManager = getSystemService(Context.LOCATION_SERVICE) as LocationManager
+        if (locationManager.isProviderEnabled(
+                LocationManager.GPS_PROVIDER)
+        ) {
+                Log.d("locationStart()","location manager Enabled")
+        } else {
+            val settingsIntent = Intent(Settings.ACTION_LOCATION_SOURCE_SETTINGS)
+            startActivity(settingsIntent)
+            Log.d("locationStart()", "not gpsEnable, startActivity")
+        }
+
+        if (ContextCompat.checkSelfPermission(this,
+                Manifest.permission.ACCESS_FINE_LOCATION) !=
+                PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(this,
+                arrayOf(Manifest.permission.ACCESS_FINE_LOCATION), 1000)
+            Log.d("locationStart", "checkSelfPermission false")
+            return
+        }
+        locationManager.requestLocationUpdates(
+            LocationManager.GPS_PROVIDER,
+            1000,
+            50f,
+            this
+        )
+    }
+
+    override fun onLocationChanged(location: Location) {
+        ShopService.instance.location = location
+        Log.d("locationchanged", "${location.latitude}, ${location.longitude}")
+    }
+
+    override fun onStatusChanged(provider: String?, status: Int, extras: Bundle?) {
+
+    }
+
+}
