+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment"
-        android:name="com.android.example.tanmen.Controller.Fragment.MainFragment"
+        android:name="jp.example.tanmen.Controller.Fragment.MainFragment"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constraintEnd_toEndOf="parent"
